Refactor difficult bcsc jwe/jwk to group into integration/bcsc folder.
